Refactored AuthConfig and split up authorization policy generator
