+++ api/v1alpha1/istiotypes/jwt_authentication.go
@@ -13,16 +13,18 @@ type Authentication struct {
 	// If omitted, the associated client registration in the application manifest is used for JWT validation.
 	SecretName *string `json:"secretName,omitempty"`
 
-	// If set to true, the original token will be kept for the upstream request. Default is true.
-	ForwardOriginalToken *bool `json:"forwardOriginalToken,omitempty"`
+	// If set to `true`, the original token will be kept for the upstream request. Defaults to `true`.
+	// +kubebuilder:default=true
+	ForwardOriginalToken bool `json:"forwardOriginalToken,omitempty"`
 
 	// Where to find the JWT in the incoming request
 	//
-	// An enum value of `header` means that the JWT is present in the `Authorization` as a Bearer Token.
+	// An enum value of `header` means that the JWT is present in the `Authorization`-header as a Bearer Token.
 	// An enum value of `cookie` means that the JWT is present as a cookie called `BearerToken`.
 	//
 	// +kubebuilder:validation:Enum=header;cookie
-	TokenLocation *string `json:"tokenLocation,omitempty"`
+	// +kubebuilder:default=header
+	TokenLocation string `json:"tokenLocation,omitempty"`
 
 	// This field specifies a list of operations to copy the claim to HTTP headers on a successfully verified token.
 	// The header specified in each operation in the list must be unique. Nested claims of type string/int/bool is supported as well.
