+++ api/v1alpha1/istiotypes/zz_generated.deepcopy.go
@@ -14,16 +14,6 @@ func (in *Authentication) DeepCopyInto(out *Authentication) {
 		*out = new(string)
 		**out = **in
 	}
-	if in.ForwardOriginalToken != nil {
-		in, out := &in.ForwardOriginalToken, &out.ForwardOriginalToken
-		*out = new(bool)
-		**out = **in
-	}
-	if in.TokenLocation != nil {
-		in, out := &in.TokenLocation, &out.TokenLocation
-		*out = new(string)
-		**out = **in
-	}
 	if in.OutputClaimToHeaders != nil {
 		in, out := &in.OutputClaimToHeaders, &out.OutputClaimToHeaders
 		*out = new([]ClaimToHeader)
