+++ pkg/reconciliation/application.go
@@ -3,6 +3,7 @@ package reconciliation
 import (
 	"context"
 	skiperatorv1alpha1 "github.com/kartverket/skiperator/api/v1alpha1"
+	"github.com/kartverket/skiperator/pkg/jwtAuth"
 	"github.com/kartverket/skiperator/pkg/log"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/rest"
@@ -14,7 +15,7 @@ type ApplicationReconciliation struct {
 
 func NewApplicationReconciliation(ctx context.Context, application *skiperatorv1alpha1.Application,
 	logger log.Logger, istioEnabled bool, restConfig *rest.Config,
-	identityConfigMap *corev1.ConfigMap, authConfigs *[]AuthConfig) *ApplicationReconciliation {
+	identityConfigMap *corev1.ConfigMap, authConfigs *jwtAuth.AuthConfigs) *ApplicationReconciliation {
 	return &ApplicationReconciliation{
 		baseReconciliation: baseReconciliation{
 			ctx:               ctx,
